local_client/requirements.txt
paho-mqtt>=2.0.0
numpy>=1.20.0
scipy>=1.7.0
matplotlib>=3.5.0
